+++ VortexEngine/src/Menus/Menus.h
@@ -81,7 +81,7 @@ private:
   // the ring menu section
   static uint8_t m_selection;
   // the time when we first opened the ringmenu
-  static uint64_t m_openTime;
+  static uint32_t m_openTime;
   // the current sub menu that is open
   static Menu *m_pCurMenu;
 };
